==> R/05_streetmap_report_large_BSB.Rmd <==
---
title: "StreetMap Analysis"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(patchwork)
# source(here::here('R/setup.R'))
source(here::here('R/munis_df.R'))
```

## Load data


```{r cars}
validated_sample_df <- 
  read_csv("../../../data/geocode/streetmap_eval/validated_large_sample_completo.csv") %>%
  mutate(geocode_result = case_when(geocode_distance <= 0 ~ "VALID",
                                  geocode_distance <= 350 ~ "GOOD",
                                  geocode_distance <= 1000 ~ "AVERAGE",
                                  TRUE ~ "BAD")) %>%
  filter(city == "Brasília")

unique(validated_sample_df$geocode_type)
geocode_type_levels <- c("PointAddress", "StreetAddress", "StreetAddressExt", "StreetName", "POI",
                         "PostalExt", "PostalLoc", "Postal", "Locality", "DistanceMarker")
```

## Data preparation


```{r pressure, echo=FALSE}
validated_count_df <- validated_sample_df %>%
  count(city, geocode_result)


city_order <- validated_count_df %>%
  pivot_wider(names_from = geocode_result, values_from = n) %>%
  ungroup() %>%
  arrange(VALID, GOOD, AVERAGE, BAD)

city_order_p  <- validated_count_df %>%
  group_by(city) %>%
  mutate(n = n / sum(n)) %>%
  pivot_wider(names_from = geocode_result, values_from = n) %>%
  ungroup() %>%
  arrange(VALID, GOOD, AVERAGE, BAD)

validated_count_df <- validated_count_df %>%
  mutate(city = factor(city, levels = city_order$city),
         geocode_result = factor(geocode_result,
                                 levels = c("VALID", "GOOD", "AVERAGE", "BAD"),
                                 labels = c("VALID", "GOOD (< 350m)", "AVERAGE (< 1km)", "BAD"))) %>%
  mutate(geocode_result = fct_rev(geocode_result))

```

Note that the `echo = FALSE` parameter was added to the code chunk to prevent printing of the R code that generated the plot.


## Resultados

### Por Município

#### Absoluto

```{r}
validated_count_df %>%
  ggplot() + geom_col(aes(x = city, y = n, fill = geocode_result)) +
  coord_flip() +
  scale_fill_brewer(palette = "Spectral", direction = 1, guide = guide_legend(reverse = TRUE)) +
  scale_y_continuous(breaks = seq(0, 25000, 2500), minor_breaks = seq(0, 25000, 1250)) +
  labs(title = "Resultado geral do geocoding, por município",
       x = NULL, y = NULL, fill = NULL,
       caption = "VALID: dentro do setor censitário\nGOOD: até 350m do setor censitário\nAVERAGE: até 1km do setor censitário\nBAD: acima de 1km do setor censitário"
  )

```

#### Relativo

```{r}
validated_count_df %>%
  group_by(city) %>%
  mutate(p = n / sum(n)) %>%
  mutate(city = factor(city, levels = city_order_p$city)) %>%
  ggplot() + geom_col(aes(x = city, y = p, fill = geocode_result)) +
  coord_flip() +
  scale_fill_brewer(palette = "Spectral", direction = 1, guide = guide_legend(reverse = TRUE)) +
  scale_y_continuous(labels = scales::percent) +
  labs(title = "Resultado geral do geocoding, por município",
       x = NULL, y = NULL, fill = NULL,
       caption = "VALID: dentro do setor censitário\nGOOD: até 350m do setor censitário\nAVERAGE: até 1km do setor censitário\nBAD: acima de 1km do setor censitário"  )


```

### Por STATUS: Match, Unmatch, Tie

```{r}
p_status_perc <- validated_sample_df %>%
  count(geocode_status, geocode_result) %>%
  group_by(geocode_status) %>%
  mutate(p = n / sum(n)) %>%
  mutate(geocode_status = factor(geocode_status,
                                 levels = c("M", "T", "U"),
                                 labels = c("MATCH", "TIE", "UNMATCH"))) %>%
  mutate(geocode_result = factor(geocode_result,
                                 levels = c("VALID", "GOOD", "AVERAGE", "BAD"),
                                 labels = c("VALID", "GOOD (< 350m)", "AVERAGE (< 1km)", "BAD"))) %>%
  mutate(geocode_result = fct_rev(geocode_result), geocode_status = fct_rev(geocode_status)) %>%
  ggplot() + geom_col(aes(x = geocode_status, y = p, fill = geocode_result)) +
  coord_flip() +
  scale_fill_brewer(palette = "Spectral", direction = 1, guide = guide_legend(reverse = TRUE)) +
  scale_y_continuous(labels = scales::percent) +
  # scale_y_continuous(breaks = seq(0, 1000, 100), minor_breaks = seq(0, 1000, 50)) +
  labs(#title = "Resultado geral do geocoding, por STATUS",
       x = NULL, y = NULL, fill = NULL
  ) +
  theme(legend.position = "bottom")

p_status_abs <- validated_sample_df %>%
  count(geocode_status, geocode_result) %>%
  group_by(geocode_status) %>%
  mutate(geocode_status = factor(geocode_status,
                                 levels = c("M", "T", "U"),
                                 labels = c("MATCH", "TIE", "UNMATCH"))) %>%
  mutate(geocode_result = factor(geocode_result,
                                 levels = c("VALID", "GOOD", "AVERAGE", "BAD"),
                                 labels = c("VALID", "GOOD (< 350m)", "AVERAGE (< 1km)", "BAD"))) %>%
  mutate(geocode_result = fct_rev(geocode_result), geocode_status = fct_rev(geocode_status)) %>%
  ggplot() + geom_col(aes(x = geocode_status, y = n, fill = geocode_result)) +
  coord_flip() +
  scale_fill_brewer(palette = "Spectral", direction = 1, guide = guide_legend(reverse = TRUE)) +
  # scale_y_continuous(breaks = seq(0, 1000, 100), minor_breaks = seq(0, 1000, 50)) +
  labs(#title = "Resultado geral do geocoding, por STATUS",
       x = NULL, y = NULL, fill = NULL
  ) +
  theme(legend.position = "bottom")

(p_status_abs + p_status_perc) +
  plot_layout(guides = "collect") & theme(legend.position = "bottom") &
  plot_annotation(title = "Resultado geral do geocoding, por STATUS",
                   caption = "VALID: dentro do setor censitário\nGOOD: até 350m do setor censitário\nAVERAGE: até 1km do setor censitário\nBAD: acima de 1km do setor censitário")
  
```

### Por ADDR_TYPE

#### Relativo

```{r}
validated_sample_df %>%
  filter(geocode_status != "U") %>%
  count(geocode_status, geocode_type, geocode_result) %>%
  group_by(geocode_status, geocode_type) %>%
  mutate(p = n / sum(n)) %>%
  mutate(geocode_type = factor(geocode_type, levels = geocode_type_levels)) %>%
  mutate(geocode_type = fct_rev(geocode_type)) %>%
  mutate(geocode_status = factor(geocode_status,
                                 levels = c("M", "T", "U"),
                                 labels = c("MATCH", "TIE", "UNMATCH"))) %>%
  mutate(geocode_result = factor(geocode_result,
                                 levels = c("VALID", "GOOD", "AVERAGE", "BAD"),
                                 labels = c("VALID", "GOOD (< 350m)", "AVERAGE (< 1km)", "BAD"))) %>%
  mutate(geocode_result = fct_rev(geocode_result)) %>%
  ggplot() + geom_col(aes(x = geocode_type, y = p, fill = geocode_result)) +
  coord_flip() +
  scale_fill_brewer(palette = "Spectral", direction = 1, guide = guide_legend(reverse = TRUE)) +
  scale_y_continuous(labels = scales::percent) +
  # scale_y_continuous(breaks = seq(0, 1000, 100), minor_breaks = seq(0, 1000, 50)) +
  labs(title = "Resultado geral do geocoding, por ADDR_TYPE",
       x = NULL, y = NULL, fill = NULL#,
       # caption = "VALID: dentro do setor censitário\nGOOD: até 350m do setor censitário\nAVERAGE: até #1km do setor censitário\nBAD: acima de 1km do setor censitário"
  ) +
  facet_wrap(~geocode_status, scales = "free", ncol = 1) +
  theme(legend.position = "right")
```

#### Absoluto

```{r}
validated_sample_df %>%
  filter(geocode_status != "U") %>%
  count(geocode_status, geocode_type, geocode_result) %>%
  group_by(geocode_status, geocode_type) %>%
  mutate(p = n / sum(n)) %>%
  mutate(geocode_type = factor(geocode_type, levels = geocode_type_levels)) %>%
  mutate(geocode_type = fct_rev(geocode_type)) %>%
  mutate(geocode_status = factor(geocode_status,
                                 levels = c("M", "T", "U"),
                                 labels = c("MATCH", "TIE", "UNMATCH"))) %>%
  mutate(geocode_result = factor(geocode_result,
                                 levels = c("VALID", "GOOD", "AVERAGE", "BAD"),
                                 labels = c("VALID", "GOOD (< 350m)", "AVERAGE (< 1km)", "BAD"))) %>%
  mutate(geocode_result = fct_rev(geocode_result)) %>%
  ggplot() + geom_col(aes(x = geocode_type, y = n, fill = geocode_result)) +
  coord_flip() +
  scale_fill_brewer(palette = "Spectral", direction = 1, guide = guide_legend(reverse = TRUE)) +
  # scale_y_percent() +
  scale_y_continuous(breaks = seq(0, 25000, 2500), minor_breaks = seq(0, 25000, 1250)) +
  labs(title = "Resultado geral do geocoding, por ADDR_TYPE",
       x = NULL, y = NULL, fill = NULL
  ) +
  facet_wrap(~geocode_status,  ncol = 1) +
  theme(legend.position = "bottom")

```

### Por Score

```{r}
plot_type_by_score <- function(types) {
  validated_sample_df %>%
  filter(geocode_status == "M") %>%
    filter(geocode_type %in% types) %>%
  # filter(str_detect(geocode_type, "Point")) %>%
  # filter(str_detect(geocode_type, "Street")) %>%
  # filter(str_detect(geocode_type, "Postal")) %>%
  # filter(geocode_type %in% c("DistanceMarker", "Locality", "POI")) %>%
  select(cnefe_id:city, geocode_status:geocode_result, -geocode_match, -geocode_rank) %>%
  mutate(score_class = round(geocode_score)) %>%
  count(geocode_type, geocode_result, score_class) %>%
  mutate(n = if_else(geocode_result %in% c("VALID", "GOOD"), n, -n),
         result_class = if_else(geocode_result %in% c("VALID", "GOOD"), "CORRECT", "WRONG")) %>%
  ggplot(aes(x=score_class, y = n, fill = result_class)) +
  geom_col() +
  geom_hline(yintercept = 0) +
  geom_vline(xintercept = 95) +
  scale_fill_brewer(palette = "Set1", direction = -1) +
  scale_x_continuous(breaks = seq(0, 100, 5), minor_breaks = 0:100) +
  facet_wrap(~geocode_type, ncol = 2, scales = "free") +
  theme(legend.position = "none")
}

```


```{r}
plot_type_by_score("PointAddress")
plot_type_by_score(c("StreetAddress", "StreetAddressExt", "StreetName"))
plot_type_by_score(c("Postal", "PostalExt", "PostalLoc"))
plot_type_by_score(c("DistanceMarker", "Locality", "POI"))

```






### Por Score Cutoff

```{r}
test_cutoffs <- function(cutoff) {
  validated_scores <- validated_sample_df %>%
    filter(geocode_status == "M") %>%
    select(cnefe_id:city, geocode_status:geocode_result, -geocode_match, -geocode_rank) %>%
    mutate(score_class = if_else(geocode_score >= cutoff, "HIGH", "LOW"),
           result_class = if_else(geocode_result %in% c("VALID", "GOOD"), "CORRECT", "WRONG")) %>%
    count(geocode_type, result_class, score_class) %>%
    pivot_wider(names_from = c(result_class, score_class), values_from = n, values_fill = 0)

  if (!("WRONG_LOW" %in% names(validated_scores))) {
    validated_scores$WRONG_LOW <- 0
  }
  if (!("WRONG_HIGH" %in% names(validated_scores))) {
    validated_scores$WRONG_HIGH <- 0
  }
  if (!("CORRECT_LOW" %in% names(validated_scores))) {
    validated_scores$CORRECT_LOW <- 0
  }
  if (!("CORRECT_HIGH" %in% names(validated_scores))) {
    validated_scores$CORRECT_HIGH <- 0
  }

  validated_scores <- validated_scores %>%
    mutate(correct = CORRECT_HIGH + WRONG_LOW,
           false_positive = CORRECT_LOW,
           false_negative = WRONG_HIGH) %>%
    mutate(score_cutoff = cutoff) %>%
    select(geocode_type, score_cutoff, correct, false_positive, false_negative)

  return(validated_scores)
}

```

```{r}
v_scores <- map_df(50:100, test_cutoffs)

```

```{r}
plot_type_by_cutoff <- function(types) {
  v_scores %>%
  filter(geocode_type %in% types ) %>%
  ggplot(aes(x=score_cutoff)) +
  geom_point(aes(y=correct)) +
  geom_path(aes(y=false_positive), colour = "blue") +
  geom_path(aes(y=false_negative), colour = "red") +
  scale_x_continuous(minor_breaks = 1:100, breaks = seq(50, 100, 5)) +
  facet_wrap(~geocode_type, scales = "free", ncol=1)

}
```


```{r}
plot_type_by_cutoff("PointAddress")
plot_type_by_cutoff(c("Postal", "PostalExt", "PostalLoc"))
plot_type_by_cutoff(c("StreetAddress", "StreetAddressExt", "StreetName"))
plot_type_by_cutoff(c("DistanceMarker", "Locality", "POI"))
```
